script-api: use more descriptive command names
Use container-usage and container-entrypoint instead of cont-help
and cont-entry.

PR: https://github.com/devexp-db/cont-lib/pull/1

* .inst-list: Distribute renamed scripts.
* bin/Makefile.inc: Build renamed scripts.
* bin/help.in: Suggest using renamed scripts.  Also, suggest using
of 'bash' instead of 'container-entrypoint' to obtain interactive
shell via 'docker exec'.  Its tough to simulate the behavior of
ENTRYPOINT statement for 'docker exec' but it is thought that
suggesting the use of 'container-entrypoint' in exec is
not conceptual and we would be widely hated.  Fix the output to be
valid markdown text, rhscl2dockerfile project uses it like that.
* configure.ac: Comment the right "usage" command, also remove the
note about proper name for '*.txt' file.  In rhscl2dockerfile
project we prefix the name by numbers to achieve correct ordering.
* share/cont-docs/README: Typofix and again mentioning right
"usage" script name.